pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.11",
    "langchain-tavily>=0.2.17",
    "langsmith>=0.7.20",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "langchain-community>=0.3.27",
    "langchainhub>=0.1.21",
    "langchain-pinecone>=0.2.11"
]
